+++ Source/SP0256/sp0256.h
@@ -100,7 +100,7 @@ void sp0256_reset();
 /*
 int sp0256_isNextSample();
 */
-int sp0256_getNextSample();
+int16_t sp0256_getNextSample();
 int sp0256_exec();
 void sp0256_setLabels( int nLabels, const char *labels[] );
 void sp0256_setDebug( int debug );
